refactor: PostsService, CommentsService 관습적 추상화 제거
