+++ src/modules/comments/comments.module.ts
@@ -1,9 +1,9 @@
 import { Module, Post } from '@nestjs/common';
-import { CommentsService } from './comment.service';
+// import { CommentsService } from './comment.service';
 import { CommentsController } from './comments.controller';
 import { CustomTypeOrmModule } from '../../config/typeorm/custom-typeorm-module';
 import { Comment } from './entities/comment.entity';
-import { CommentsServiceImpl } from './comments-impl.service';
+import { CommentsService } from './comments.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from '../users/entities/user.entity';
 import { CommentsRepository } from './comments.repository';
@@ -20,6 +20,6 @@ import { PostsRepository } from '../posts/posts.repository';
     ]),
   ],
   controllers: [CommentsController],
-  providers: [{ provide: CommentsService, useClass: CommentsServiceImpl }],
+  providers: [CommentsService],
 })
 export class CommentsModule {}
