+++ src/test/controller/comments.controller.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { CommentsController } from '../../modules/comments/comments.controller';
-import { CommentsService } from '../../modules/comments/comment.service';
-import { CommentsServiceImpl } from '../../modules/comments/comments-impl.service';
+// import { CommentsService } from '../../modules/comments/comment.service';
+import { CommentsService } from '../../modules/comments/comments.service';
 import { CommentsRepository } from '../../modules/comments/comments.repository';
 import { UsersRepository } from '../../modules/users/users.repository';
 import { PostsRepository } from '../../modules/posts/posts.repository';
@@ -13,7 +13,7 @@ describe('CommentsController', () => {
     const module: TestingModule = await Test.createTestingModule({
       controllers: [CommentsController],
       providers: [
-        { provide: CommentsService, useClass: CommentsServiceImpl },
+        CommentsService,
         CommentsRepository,
         UsersRepository,
         PostsRepository,
