+++ src/test/controller/posts.controller.spec.ts
@@ -1,8 +1,8 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { PostsController } from '../../modules/posts/posts.controller';
-import { PostsServiceImp } from '../../modules/posts/posts-impl.service';
-import { PostsRepository } from '../../modules/posts/posts.repository';
 import { PostsService } from '../../modules/posts/posts.service';
+import { PostsRepository } from '../../modules/posts/posts.repository';
+// import { PostsService } from '../../modules/posts/posts.service';
 import { HttpStatus } from '@nestjs/common';
 import { UsersRepository } from '../../modules/users/users.repository';
 import { CategorysRepository } from '../../modules/categorys/categorys.repository';
@@ -40,10 +40,7 @@ describe('PostsController', () => {
         PostsRepository,
         UsersRepository,
         CategorysRepository,
-        {
-          provide: PostsService,
-          useClass: PostsServiceImp,
-        },
+        PostsService,
         { provide: DataSource, useValue: mockDataSource },
         {
           provide: getRepositoryToken(PostLike),
